Drop BaseToolBox: use AbstractToolBox as base class instead
