+++ lib/galaxy/tool_util/toolbox/__init__.py
@@ -1,9 +1,6 @@
 """API for this module containing functionality related to the toolbox."""
 
-from .base import (
-    AbstractToolBox,
-    BaseGalaxyToolBox,
-)
+from .base import AbstractToolBox
 from .panel import (
     panel_item_types,
     ToolSection,
@@ -12,7 +9,6 @@ from .panel import (
 
 __all__ = (
     "AbstractToolBox",
-    "BaseGalaxyToolBox",
     "panel_item_types",
     "ToolSection",
     "ToolSectionLabel",
