+++ lib/galaxy/tool_util/toolbox/base.py
@@ -1410,72 +1410,3 @@ def _filter_for_panel(item, item_type, filters, context):
                 return copy
 
     return None
-
-
-class BaseGalaxyToolBox(AbstractToolBox):
-    """
-    Extend the AbstractToolBox with more Galaxy tooling-specific
-    functionality. Adds dependencies on dependency resolution and
-    tool loading modules, that an abstract description of panels
-    shouldn't really depend on.
-    """
-
-    def __init__(
-        self,
-        config_filenames,
-        tool_root_dir,
-        app,
-        view_sources=None,
-        default_panel_view=None,
-        save_integrated_tool_panel=True,
-    ):
-        super().__init__(
-            config_filenames, tool_root_dir, app, view_sources, default_panel_view, save_integrated_tool_panel
-        )
-        old_toolbox = getattr(app, "toolbox", None)
-        if old_toolbox:
-            self.dependency_manager = old_toolbox.dependency_manager
-        else:
-            self._init_dependency_manager()
-
-    @property
-    def sa_session(self):
-        """
-        Returns a SQLAlchemy session
-        """
-        return self.app.model.context
-
-    def _looks_like_a_tool(self, path):
-        return looks_like_a_tool(path, enable_beta_formats=getattr(self.app.config, "enable_beta_tool_formats", False))
-
-    def _init_dependency_manager(self):
-        use_tool_dependency_resolution = getattr(self.app, "use_tool_dependency_resolution", True)
-        if not use_tool_dependency_resolution:
-            self.dependency_manager = NullDependencyManager()
-            return
-        app_config_dict = self.app.config.config_dict
-        conf_file = app_config_dict.get("dependency_resolvers_config_file")
-        default_tool_dependency_dir = os.path.join(
-            self.app.config.data_dir, self.app.config.schema.defaults["tool_dependency_dir"]
-        )
-        self.dependency_manager = build_dependency_manager(
-            app_config_dict=app_config_dict,
-            conf_file=conf_file,
-            default_tool_dependency_dir=default_tool_dependency_dir,
-        )
-
-    def reload_dependency_manager(self):
-        self._init_dependency_manager()
-
-    def load_builtin_converters(self):
-        id = "builtin_converters"
-        section = ToolSection({"name": "Built-in Converters", "id": id})
-        self._tool_panel[id] = section
-
-        converters = self.app.datatypes_registry.datatype_converters
-        for source, targets in converters.items():
-            for target, tool in targets.items():
-                tool.name = f"{source}-to-{target}"
-                tool.description = "converter"
-                tool.hidden = False
-                section.elems.append_tool(tool)
